pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "scheduler-tools"
version = "0.0.1"
description = "Basic tools for automating some PBS work. In progress, and potentially unsafe."
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = [
    "apscheduler>=3.11.0",
    "paramiko>=3.5.1",
    "python-daemon>=3.1.2",
    "regex>=2024.11.6",
    "rich>=14.0.0",
]

[project.optional-dependencies]
ui = [
  "textual>=3.2.0",
]

[project.scripts]
check-status = "schedtools.interfaces.check_status:check_status"
clear-logs = "schedtools.interfaces.clear_logs:clear_logs"
convert-jobscripts = "schedtools.interfaces.convert_jobscripts:convert_jobscripts"
create-smtp-credentials = "schedtools.smtp:create_credentials"
delete-duplicate-jobs = "schedtools.interfaces.delete_duplicate_jobs:delete_duplicate_jobs"
queue-manager = "schedtools.interfaces.queue_manager.__main__:main"
remote-command = "schedtools.interfaces.remote_command:remote_command"
remove-service = "schedtools.interfaces.remove_service:remove_service"
rerun = "schedtools.interfaces.rerun:rerun"
schedtools-help = "schedtools.interfaces.help:help"
storage-tracker = "schedtools.interfaces.storage_tracker:storage_tracker"
update-job-state = "schedtools.interfaces.update_job_state:update_job_state"


[dependency-groups]
dev = [
  "pytest>=6.0",
  "pytest-cov>=4.0",
]


[tool.pytest.ini_options]
addopts = "--cov=schedtools"
markers = [
    "nohidecreds: Do not hide SMTP credentials (if they exist) for this test.",
]
testpaths = [
    "test",
]
